+++ src/domain/map/MapControl.ts
@@ -1,51 +1,43 @@
+import { createControlComponent } from "@react-leaflet/core";
 import L from "leaflet";
 import { ReactElement } from "react";
-import ReactDOM from "react-dom";
-import {
-  MapControl as RLMapControl,
-  withLeaflet,
-  MapControlProps,
-} from "react-leaflet";
+import { createRoot } from "react-dom/client";
 
-type Props = MapControlProps & {
+type Props = {
   handleClick: (e: Event) => void;
   className?: string;
   children: ReactElement;
+  position?: L.ControlPosition;
 };
 
-class MapControl extends RLMapControl<Props> {
-  // note we're extending MapControl from react-leaflet, not Component from react
-  constructor(props: Props) {
-    super(props);
-    this.handleClick = this.handleClick.bind(this);
-  }
+const createMapControl = (props: Props) => {
+  const { className, children, handleClick, position } = props;
 
-  handleClick(event: Event) {
-    L.DomEvent.stopPropagation(event);
+  const control = new L.Control({
+    position: position || "bottomright",
+  });
 
-    this.props.handleClick && this.props.handleClick(event);
-  }
+  control.onAdd = () => {
+    const div = L.DomUtil.create("div", `custom-control ${className}`);
+    const link = L.DomUtil.create("button", "custom-control-button", div);
 
-  createLeafletElement({ className, children }: Props) {
-    const control = new L.Control({
-      position: this.props.position || "bottomright",
-    });
+    L.DomEvent.on(link, "mousedown dblclick", L.DomEvent.stopPropagation)
+      .on(link, "click", L.DomEvent.stop)
+      .on(link, "click", (event: Event) => {
+        L.DomEvent.stopPropagation(event);
+        handleClick && handleClick(event);
+      });
 
-    control.onAdd = () => {
-      const div = L.DomUtil.create("div", `custom-control ${className}`);
-      const link = L.DomUtil.create("button", "custom-control-button", div);
+    // Render React element into the control
+    const root = createRoot(link);
+    root.render(children);
 
-      L.DomEvent.on(link, "mousedown dblclick", L.DomEvent.stopPropagation)
-        .on(link, "click", L.DomEvent.stop)
-        .on(link, "click", this.handleClick, control);
+    return div;
+  };
 
-      ReactDOM.render(children, link);
+  return control;
+};
 
-      return div;
-    };
+const MapControl = createControlComponent(createMapControl);
 
-    return control;
-  }
-}
-
-export default withLeaflet(MapControl);
+export default MapControl;