+++ src/domain/map/mapConstants.ts
@@ -1,20 +1,6 @@
 import { LatLngBoundsLiteral } from "leaflet";
-import { MutableRefObject } from "react";
-import { Map } from "react-leaflet";
 
 import { normalizeActionName } from "../utils";
-
-export const MAP_URL = {
-  fi: "https://tiles.hel.ninja/styles/hel-osm-light/{z}/{x}/{y}.png",
-  sv: "https://tiles.hel.ninja/styles/hel-osm-light/{z}/{x}/{y}@sv.png",
-  en: "https://tiles.hel.ninja/styles/hel-osm-light/{z}/{x}/{y}@en.png",
-};
-
-export const MAP_RETINA_URL = {
-  fi: MAP_URL.fi.replace(".png", "@3x.png"),
-  sv: MAP_URL.sv.replace("@sv", "@3x@sv"),
-  en: MAP_URL.en.replace("@en", "@3x@en"),
-};
 
 export const DEFAULT_ZOOM = 12;
 
@@ -33,5 +19,3 @@ export const mapActions = {
   SET_LOCATION: normalizeActionName("map/SET_LOCATION"),
   RECEIVE_ADDRESS: normalizeActionName("map/RECEIVE_ADDRESS"),
 };
-
-export type MapRef = MutableRefObject<Map | null> | ((ref: Map) => void);